[OW-001] feat: ktlint, libs versions, modularization, editorconfig, buildlogic
