+++ .editorconfig
@@ -1,0 +1,12 @@
+root = true
+
+
+[*.{kt,kts}]
+max_line_length = off
+ij_kotlin_keep_blank_lines_in_code = 0
+ij_kotlin_keep_blank_lines_in_declarations = 1
+ij_kotlin_blank_lines_after_class_header = 1
+ij_kotlin_keep_blank_lines_before_right_brace = 0
+ktlint_standard_no-wildcard-imports = disabled
+
+
