+++ .github/workflows/tests.yml
@@ -1,0 +1,38 @@
+name: Tests
+
+on:
+  push:
+    branches: [ master ]
+  pull_request:
+    branches: [ master ]
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout apollo
+        uses: actions/checkout@v3
+      - name: Set up JDK 17
+        uses: actions/setup-java@v3
+        with:
+          java-version: 17
+          distribution: 'temurin'
+          cache: gradle
+
+      - name: Grant execute permission for gradlew
+        run: chmod +x gradlew
+
+      - name: Lint
+        run: ./gradlew ktlintCheck
+
+      - name: Gradle Test
+        run: ./gradlew test --rerun-tasks --stacktrace
+
+      - name: Generate failure report if required
+        if: failure()
+        uses: actions/upload-artifact@v3
+        with:
+          name: unit-test-results
+          path: common/build/reports/tests/testDebugUnitTest/
+
