+++ backend/src/main/kotlin/dev/pango/apollo/backend/modules/userauth/data/dto/user/AddUserDTO.kt
@@ -1,0 +1,10 @@
+package dev.pango.apollo.backend.modules.userauth.data.dto.user
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class AddUserDTO(
+    val firstName: String,
+    val lastName: String,
+    val email: String,
+)
