+++ backend/src/main/kotlin/dev/pango/apollo/backend/modules/userauth/data/dto/user/SearchUserDTO.kt
@@ -1,0 +1,5 @@
+package dev.pango.apollo.backend.modules.userauth.data.dto.user
+
+data class SearchUserDTO(
+    val id: Int,
+)
