+++ backend/src/main/kotlin/dev/pango/apollo/backend/modules/userauth/data/dto/user/UpdateUserDTO.kt
@@ -1,0 +1,8 @@
+package dev.pango.apollo.backend.modules.userauth.data.dto.user
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class UpdateUserDTO(
+    val firstName: String,
+)
