+++ backend/src/main/kotlin/dev/pango/apollo/backend/modules/userauth/di/UserAuthModule.kt
@@ -1,0 +1,12 @@
+package dev.pango.apollo.backend.modules.userauth.di
+
+import dev.pango.apollo.backend.modules.userauth.data.repository.UserRepositoryImpl
+import dev.pango.apollo.backend.modules.userauth.domain.repository.UserRepository
+import org.koin.dsl.module
+
+val userAuthModule =
+    module {
+        single<UserRepository> {
+            UserRepositoryImpl()
+        }
+    }
