+++ backend/src/main/kotlin/dev/pango/apollo/backend/plugins/Routing.kt
@@ -1,0 +1,9 @@
+package dev.pango.apollo.backend.plugins
+
+import io.ktor.server.application.*
+import io.ktor.server.routing.*
+
+fun Application.configureRouting() {
+    routing {
+    }
+}
