+++ gradle.properties
@@ -1,8 +1,1 @@
-ktor_version=2.3.7
-kotlin_version=1.9.21
-logback_version=1.4.11
 kotlin.code.style=official
-exposed_version = 0.41.1
-h2_version = 2.1.21
-postgresql_driver_version=42.3.1
-ktorm_version=3.6.0
