+++ settings.gradle.kts
@@ -1,1 +1,19 @@
-rootProject.name = "dev.pango.osar-back"
+
+pluginManagement {
+    includeBuild("buildlogic")
+    repositories {
+        mavenCentral()
+        google()
+        gradlePluginPortal()
+    }
+}
+
+dependencyResolutionManagement {
+    repositories {
+        mavenCentral()
+        google()
+    }
+}
+
+rootProject.name = "apollo"
+include(":backend")
