--- src/main/kotlin/dev/pango/dao/FileService.kt
@@ -1,31 +1,0 @@
-package dev.pango.dao
-
-import dev.pango.models.*
-import org.ktorm.database.Database
-import org.ktorm.dsl.*
-import org.ktorm.entity.*
-import java.util.UUID
-
-class FileService {
-    private val database = Database.connect(
-        url = "jdbc:postgresql://localhost:5432/postgres",
-        driver = "org.postgresql.Driver",
-        user = "postgres",
-        password = "postgres"
-    )
-
-    fun createFile(filesRequest: FilesRequest): Boolean {
-        val newFile = File {
-            id = filesRequest.id
-            path = filesRequest.path
-        }
-        val affectedRecordsNumber =
-            database.sequenceOf(Files)
-                .add(newFile)
-        return affectedRecordsNumber == 1
-    }
-
-    fun findFileById(fileId: String): File =
-        database.sequenceOf(Files).single { it.id eq UUID.fromString(fileId) }
-
-}
