--- src/main/kotlin/dev/pango/modules/sharedkernel/infraestructure/persistence/config.kt
@@ -1,10 +1,0 @@
-package dev.pango.modules.sharedkernel.infraestructure.persistence
-
-import org.ktorm.database.Database
-
- val database = Database.connect(
-    url = "jdbc:postgresql://localhost:5432/postgres",
-    driver = "org.postgresql.Driver",
-    user = "postgres",
-    password = "postgres"
-)
