--- src/main/kotlin/dev/pango/modules/sharedkernel/presentation/apirest/safeApiCall.kt
@@ -1,4 +1,0 @@
-package dev.pango.modules.sharedkernel.presentation.apirest
-
-import arrow.core.Either
-import dev.pango.modules.sharedkernel.domain.failure.Failure
