--- src/main/kotlin/dev/pango/modules/userauth/data/dto/user/AddUserDTO.kt
@@ -1,10 +1,0 @@
-package dev.pango.modules.userauth.data.dto.user
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class AddUserDTO(
-    val firstName: String,
-    val lastName: String,
-    val email: String,
-)
