--- src/main/kotlin/dev/pango/modules/userauth/data/dto/user/SearchUserDTO.kt
@@ -1,5 +1,0 @@
-package dev.pango.modules.userauth.data.dto.user
-
-data class SearchUserDTO(
-    val id: Int,
-)
