--- src/main/kotlin/dev/pango/modules/userauth/data/dto/user/UpdateUserDTO.kt
@@ -1,9 +1,0 @@
-package dev.pango.modules.userauth.data.dto.user
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class UpdateUserDTO(
-    val firstName: String,
-)
-
