--- src/main/kotlin/dev/pango/plugins/Routing.kt
@@ -1,11 +1,0 @@
-package dev.pango.plugins
-
-import dev.pango.routes.*
-import io.ktor.server.application.*
-import io.ktor.server.routing.*
-import io.ktor.server.plugins.swagger.*
-
-fun Application.configureRouting() {
-    routing {
-    }
-}
