--- src/main/resources/application.conf
@@ -1,21 +1,0 @@
-ktor {
-    deployment {
-        port = 8080
-        port = ${?PORT}
-    }
-    application {
-        modules = [ dev.pango.ApplicationKt.module ]
-    }
-}
-jwt {
-    secret = "secret"
-    issuer = "http://0.0.0.0:8080/"
-    audience = "http://0.0.0.0:8080/hello"
-    realm = "Access to 'hello'"
-}
-postgres {
-    url = "Your URL"
-    user = "Your USERNAME"
-    password = "Your PASSWORD"
-}
-
